Added Static Friction Distortion
This effect is based on static friction - whenever the incoming signal peaks (over a threshold) it remains at that peak value for a while before the change becomes great enough to drag it back down. This is an extension of the updated acceleration limiter, and it also has a regular friction value that simply reduces speed constantly.
As usual, it sounds like some kind of archaic recording distortion, though it has some mildly novel effects. For instance, it tends to heavily emphasize bass notes or kick drums over the rest of the signal briefly. Additionally, this is a very asymmetrical distortion, and will have a different sound quality than other distortions that are centered over peaks.